Create pt3.java
3. Design and implement the class Day that implements the day of the week in a program. The class Day should store the day, such as Sun for Sunday. The program should be able to perform the following operations on an object of type Day:

Set the day.
Print the day.
Return the day.
Return the next day.
Return the previous day.
Calculate and return the day by adding certain days to the current day. For example, if the current day is Monday and we add four days, the day to be returned is Friday.
Add the appropriate constructors.
Write the definitions of the methods to implement the operations for the class Day, as defined in a through g.
Write a program to test various operations on the class Day.